flexalloc: Pass fs to notfs_geo_slab_lb_off
Now we can set the slab lb off to zero if we have a MD device. Also updated the
alignment to a zone boundary to only happen if there is a problem. This should
not occur since we make the checks ahead of time. TODO add assertion in this
function for slabs unaligned to zone boundaries.

Signed-off-by: Adam Manzanares <a.manzanares@samsung.com>