flexalloc: Add striped object support
Adds support for striping writes across objects at the pool level. The function
fla_pool_set_strp takes an argument of the number of stripes as well as the
strp sz to set for a pool. The number of stripes sets the number of objects
that will be allocated during object creation time. The strp sz is the unit of
IO that will be sent to each object.

Signed-off-by: Adam Manzanares <a.manzanares@samsung.com>