recv - error-check for empty messages
Set CMD to FLA_MSG_CMD_NULL, indicating a non-command - only if a
new message is actually received will the return value be 0.
Before, recv would return 0 and the buffer would reflect the prior
received CMD.
This catches cases where the client is disconnected, for example
due to sending a malformed/unrecognized command.

Signed-off-by: Jesper Wendel Devantier <j.devantier@samsung.com>