SlabCache: Adjust counter considering num objects
When adjusting the object counter in each slab we adjusted only one
object at a time. Now we need to adjusted it by the number of objects in
the stripe to keep correct count.

*/src/flexalloc_mm.c :
(fla_slab_next_available_obj) : When we get an object back adjust
counter by number of objects in stripe
(fla_object_destroy) : Likewise

Signed-off-by: Joel Granados <j.granados@samsung.com>