ZNS: fla_object_close -> fla_object_seal
What we are doing with the previously named fla_object_close is to
transition the ZNS zone into FULL state which will disallow writes. ZNS
specifies that the only state transition from FULL is to emtpy, so it can't
really be re-opened for further writes.
The word "close" when referring to storage implicitly implies that there
could be a subsequent open and more writes can be performed. To avoid
confusion we remove "closed" from the name of this function and replace it
with "seal" which reflects more what is actually happening.

Signed-off-by: Joel Granados <j.granados@samsung.com>