freelist: Add generic search function
We leverage the effective way we traverse our free list to execute a
generic function for every non-free element. The intended way to
introduce the new fla_flist_search_wfunc function is:

  int ret;
  uint32_t found;
  ret = fla_flist_search_wfunc(flist, type, &found, func)

Where "found" will have the amount of times the "func" function found
something in the freelist. The function also accepts additional variadic
arguments so it can pass all the necessary data to the func call.

Signed-off-by: Joel Granados <j.granados@samsung.com>