DP: Add data placement indirection
For now data placement (DP) is confined to changing the xnvme context on the
write path. We define the data placement policy at fs_init by setting a
pointer to DP specific struct and a set of DP specific callbacks used
mainly to change the context on the write path.

This patch adds a noop DP engine which basically encapsulates the previous
behavior of doing nothing with respect to data placement. It is worth
noting that the DP callback needs to be set in xne_io everytime that it
gets created.

Signed-off-by: Joel Granados <j.granados@samsung.com>