DP: Add FDP data placement engine
Using the infrastructure for the Data Placement indirection (DP) we add
FDP. Given that we want to experiment having different FDP policies based
on slabs, pools and writes, we create several context callbacks that will
behave differently depending on if it is in the pool, slab or object.

This patch introduces a FDP struct that holds in memory fdp specific data
like what pid is related to what object. The memory is realeased when the
dp_fini is called.

Signed-off-by: Joel Granados <j.granados@samsung.com>