flan: Modify the env_uri parsing function
In the initialization string passed through env_uri to rocksDB we used to
include DEV:STRP_TOTAL_SIZE:MD_DEV. We change it so it now expects the
following:
  flan:DEV:MD_DEV:SSTableSize:OBJ_NBYTES:STRP_NOBJS:STRP_NBYTES
where DEV is the main device, MD_DEV is the metadata device and can be "".
SSTableSize is the rocksDB sstable size in bytes. Flexalloc Object size in
bytes (OBJ_NBYTES), STRP_NOBJS is the number of fla objects in a striped
object and STRP_NBYTES is the number of bytes in a stripped section within
the fla object.

Signed-off-by: Joel Granados <j.granados@samsung.com>